+++ pkg/linstorhelper/client_test.go
@@ -215,7 +215,7 @@ func TestNewClientForCluster(t *testing.T) {
 				require.NoError(t, err)
 
 				// need to use go-cmp here, as that can handle the embedded x509.CertPool comparison.
-				diff := cmp.Diff(*expected, actual.Client,
+				diff := cmp.Diff(expected, actual.Client,
 					// Compare all unexported fields, too
 					cmp.Exporter(func(r reflect.Type) bool {
 						return true
@@ -390,7 +390,7 @@ func TestCreateOrUpdateNode(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			t.Parallel()
 
-			lc := linstorhelper.Client{Client: lapi.Client{
+			lc := linstorhelper.Client{Client: &lapi.Client{
 				Nodes: test.setupCalls(t),
 			}}